(chore) test: deduplicate JNA/FFM loading tests
Extract shared loading failure-mode tests into a
Pcre2LoadingContractTest contract interface in the lib testFixtures,
following the existing contract test pattern. JNA and FFM
Pcre2LoadingTests now implement the contract interface and provide
only the backend-specific createBackend factory method.

Closes #329
